Fix use of get_database_nodes function
References to the value returned by this function were using the static
database ip attribute for each node returned. With this change the IP of
the nodes returned by the function are used.

Where the ['contrail']['database']['ip'] attribute was being used in
isolation the current node's IP address is being used instead. This will
be still be the database ndoe as its only used on the database node.