Get contrail controller node ip address from Ohai
This change replaces references to these attributes,
* ['contrail']['controller']['ip']
* ['contrail']['controller']['hostname']
with these ohai attributes.
* ['ipaddress']
* ['hostname']

Since the Ohai attributes are dynamically created on each run it means
we don't have to know these values before we run the recipes.

A method to search for the Contrail Controller IP has also been added to
the utils library for future use.